Removed bugged package pkg-resources==0.0.0
As mentioned in these discussion, pkg-resources is a bug derived from executing `pip freeze` on Ubuntu. It breaks dependency installations on MacOSX `pip3 install -r requirements.txt` with the following error `ERROR: Could not find a version that satisfies the requirement pkg-resources==0.0.0 (from -r requirements.txt (line 9)) (from versions: none)`

This file is not required and is an artefact of an old issue, hence removed.